Extract shared retrieval core from tools.py (no behavior change)
Factor the retrieval/fusion plumbing shared by lcm_grep into a new
retrieval_core.py so the forthcoming lcm_recall tool reuses one ranking
engine instead of duplicating logic. Pure move — every tool contract,
JSON shape, degrade path, and error string is byte-identical.

Moved: RRF fusion (rrf_fuse), the KNN invocation (run_knn) and node
hydration, confidence scoring, deadline-error shaping, semantic
conversation-scope resolution, and the FTS message/summary hit shapers.
Guards (role/time/conversation/scope/content_scope), the bounded-worker
deadline machinery, and provider-resolution/query-embedding stay in
tools.py: their module-level names are monkeypatched through the tools
namespace and the worker-slot default binding is lexical to that module.

VectorStore is injected into run_knn so callers keep resolving it through
tools.VectorStore (preserves test monkeypatch semantics).
